Move settings to the integration, call it just Settings
